+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/enc/JweService.java
@@ -10,18 +10,19 @@ import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerCredentialResponseE
 import ch.admin.bj.swiyu.issuer.domain.openid.metadata.IssuerMetadata;
 import ch.admin.bj.swiyu.jweutil.JweUtil;
 import ch.admin.bj.swiyu.jweutil.JweUtilException;
-import com.nimbusds.jose.*;
+import com.nimbusds.jose.JWEAlgorithm;
+import com.nimbusds.jose.JWEHeader;
+import com.nimbusds.jose.JWEObject;
 import com.nimbusds.jose.jwk.JWK;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Map;
 import java.text.ParseException;
+import java.util.Map;
 
 import static ch.admin.bj.swiyu.issuer.common.exception.CredentialRequestError.INVALID_ENCRYPTION_PARAMETERS;
 
@@ -38,6 +39,7 @@ public class JweService {
     private final IssuerMetadata issuerMetadata;
     private final EncryptionKeyService encryptionKeyService;
 
+
     /**
      * Overriding bean issuer metadata encryption options with supported values.
      */
@@ -53,6 +55,7 @@ public class JweService {
         issuerMetadata.setResponseEncryption(IssuerCredentialResponseEncryption.builder()
                 .encRequired(applicationProperties.isEncryptionEnforce())
                 .build());
+
         return issuerMetadata;
     }